requirements.txt
streamlit
openai
pinecone
langchain
langchain-community
python-dotenv
